String [#237407]

+++ src/main/java/ru/job4j/condition/Greeting.java
@@ -5,7 +5,7 @@ public class Greeting {
         String idea = "i like Java! ";
         System.out.println(idea);
         idea += "But im a newbie. ";
-        short year = 2022;
+        int year = 2022;
         idea += year;
         System.out.println(idea);
     }
